requirements.txt
# Groww API SDK
growwapi

# Environment variable loading
python-dotenv

# TOTP authentication (optional but recommended)
pyotp

# For instrument dataframe handling
pandas
